fix: update parameter name from torch_dtype to dtype in model initialization

+++ whisper_transcribe_py/audio_transcriber.py
@@ -296,7 +296,7 @@ class WhisperTranscriber:
         device, torch_dtype = _get_device_and_dtype()
 
         model = AutoModelForSpeechSeq2Seq.from_pretrained(
-            model_id, torch_dtype=torch_dtype, low_cpu_mem_usage=True, use_safetensors=True
+            model_id, dtype=torch_dtype, low_cpu_mem_usage=True, use_safetensors=True
         )
         model.to(device)
 
@@ -307,7 +307,7 @@ class WhisperTranscriber:
             model=model,
             tokenizer=processor.tokenizer,
             feature_extractor=processor.feature_extractor,
-            torch_dtype=torch_dtype,
+            dtype=torch_dtype,
             device=device,
         )
